requirements.txt
asgiref
async-generator
attrs
beautifulsoup4
certifi
charset-normalizer
Django
django-crispy-forms
django-environ
django-froala-editor
django-widget-tweaks
djangorestframework
gunicorn
Pillow
psycopg2-binary
requests
selenium
